VLN_win/requirements.txt
airsim==1.8.1
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv>=1.0.0
